+++ src/main/java/com/alper/worldcup/dao/FinalPredictionAuditRepository.java
@@ -6,6 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FinalPredictionAuditRepository extends JpaRepository<FinalPredictionAudit, Integer> {
 
+    long countByAction(com.alper.worldcup.entity.PredictionAuditAction action);
+
     List<FinalPredictionAudit> findAllByOrderByRecordedAtDesc();
 
     List<FinalPredictionAudit> findByUsernameOrderByRecordedAtDesc(String username);